fix<ScriptableObj>: change earn price from double to long

+++ Assets/Resources/ScriptableObjects/GraphicCardItem.cs
@@ -12,8 +12,8 @@ public class GraphicCardItem : ScriptableObject
     public string Id;
     public Sprite Icon;
     public bool IsLocked;
-    public double PerSecondEarn;
+    public long PerSecondEarn;
     public long Price;
-    public double PerSecondLoseVolt;
+    public long PerSecondLoseVolt;
     public int Quantity;
 }
